Cipher: Now encryption and decryption is done with a Key object
Now to encrypt and decrypt we use a Key object. This class is just a
wrapper for an Object that has the key, this way the key can later be
casted into the needed class.

This makes the code easier to expand in the future, in case we add a
new Cipher that don't takes an Integer as the encryption key (e.g.
Vigenère cipher, takes a keyword).