Deciphering: First approach into deciphering ciphed text
In this commit I added the analysis package, wich contains useful tools
for deciphering text ciphed with the implemented ciphers (A1Z26, Atbash
and Caesar).

The Decrypter class recieves an encrypted text and returns a list of
DecryptGuess sorted by error. Right now the decryption is done with a
Brute-Force attack, testing every cipher. Also only one step into
decryption is taken into account, meaning that if a text is encrypted
multiple times it won't be possible to find the original text (This is
on the TODO list).

For making the deciphering process easier i've added the abstract method
decryptWithoutKey(String, Analyzer) on the Cipher class, this method
returns a list of DecryptGuess.

The DecryptGuess class acts as a chain of nodes, each node being a step
in the decryption process. This class consists in:
    - a List of cipher names, first one being the first one used on the
      ciphed text
    - a List of Keys, same order as the list of cipher names
    - the resulting decrypted text
    - the final error, the lower the error the bigger the probability
      that the resulting decrypted text is from the language specified
      in the Config class

As a first Analyzer i've implemented a Frequency Analysis class, this
just calculates the frequency of the letters in the given text and
compares it to the frequency stored on the Config class. The comparation
is just the Euclidean distance squared between the two frequencies, I
treat them as two vectors.