Renamed utils/Color.java to utils/Colour.java
I changed the name of the class because NetBeans was underlining the
comments that had "color" in them as if it was misspelled. I know that
Color is correct, it's mostly used in the US. I also know that NetBeans
has the option to add a word to the project dictionary, but i prefer to
just rename the class.