Change implementation of --no-color flag to be more intuitive
Now the names of the arguments related to the colour display are more
intuitive, if noColour_ is true the colour is not used.

Also changed some more names from Color to Colour, this should be the
last time I have to change something related to the colour, probably.

Also changed the flag from run.bat to the new one.