Now the Config class is needs a Namespace in the constructor
Now the Config class also holds a Namespace as an attribute (the same
namespace that the method ArgumentParser.parseArgs() returns).

It's done this way because now will be easier and safer (from Config
you only can get values not put them) to acces all the configuration
values stored in the Namespace, it's only needed to pass the Config
object to other classes.