Ciphers: A1Z26 solved problem when parsing  big numbers
Before the method Integer.valueOf() was called directly after checking
if the value was an Integer, this can cause some problem if the number
to convert is bigger than Integer.MAX_VALUE.

There was also the problem that the number parsed could be bigger than
the alphabet lenght and then throw an ArrayIndexOutOfBoundsException.

Solved both of this problems.