Ciphers: super-class no longer has static attr. NAME and modified decryption without key
Removed the static variable NAME from Cipher because it was redundant,
also made the method getName() abstract.

Now decryptWithoutKey() takes a DecryptGuess as an input. If it's null
this means is the first one so we just create a new one, in case it's
not, that means we have to add a step. In case that the text once
dectypted is the same as before decryption we return an empty list.