Decrypted: Added support for multiple cipher decryption
Now we can decrypt texts that had been ciphed multiple times with
different ciphers. :^)

This makes the project run a little bit slower because it has to
calculate much more possibilites.

Also the code is a little bit messy in decrypt(), I should make it more
clear and add comments.