DecryptGuess: The cipher names now have right padding
Now the cipher names on the decoded file all have the same length,
MAX_CIPHER_NAMES_WIDTH = 27. The value of this variable is the sum of all the
cipher names, example:
    [A1Z26][Atbash][Caesar 20] -> 26
    [Binary][Atbash][Caesar 20] -> 27
    [Binary][A1Z26][Atbash][Caesar 20] -> 34

As it can be seen, although the max value is 34 I've chosen 27 as the
max value, this is because most of the time the correct result will only
have 3 ciphers at max.

I should calculate this value dynamically, and not set it as a constant,
in case I add more ciphers to the program.