Config: Arguments and configuration values now have prefixes
Added prefixes to all the configuration values, this way is easier to
separate the values into "themes". For example separate all the io
configuration from the menu or decryption/encryption one.

The prefixes are separated by a dot ".", an example:
    io.debug = True
This would enable debug mode.

TODO: change the config file from yaml to this dot separated one, create
a file parser and store all in a Namespace.